charge pump Project
Developed a charge pump circuit as part of a lab project, focusing on design, simulation, and testing to ensure efficiency and reliability.
It is addressing the need for multiple voltage sources in VLSI components. It involved the implementation of three different charge pump circuits, each featuring a voltage stabilization control system based on a switched comparator protected against offset errors. The design targeted a 180nm process, including analog simulations and circuit layout.